Use `gensym` for spy protocol args
The previous version used the names defined in the protocol definition
for the spy args, this breaks when using destructuring within the
protocol definition and fails to pass the args through correctly. We
now use gensym to generate custom arg names within the protocol.

Test destructure within definition but not implementation